ols_core/requirements.txt
yaml~=0.2.5
pyyaml~=6.0
setuptools~=61.2.0
dataclasses~=0.8
twisted~=22.2.0
greenlet~=1.1.1
jinja2~=2.11.3
urllib3~=1.26.9
requests~=2.27.1
tabulate~=0.8.9
scipy~=1.7.3
python-dateutil~=2.8.2
six~=1.16.0
pyjwt~=2.1.0
packaging~=21.3
tenacity~=8.0.1
pygments~=2.11.2
cython~=0.29.28
scikit-learn~=1.2.1
pandas~=1.4.2
toml~=0.10.2
numpy~=1.21.5
filelock~=3.6.0
cachetools~=4.2.2
pyarrow~=10.0.0
cloudpickle~=2.0.0
click~=8.0.4
colorama~=0.4.4
pillow~=9.0.1
sortedcontainers~=2.4.0
tqdm~=4.64.0
docutils~=0.17.1
sphinx~=4.4.0
matplotlib~=3.5.1
seaborn~=0.11.2
flask~=1.1.2
h5py~=3.6.0
nltk~=3.7
transformers~=4.24.0
networkx~=2.7.1
psutil~=5.8.0
certifi~=2023.7.22
pydantic~=1.10.9
attrs~=21.4.0
distributed~=2022.2.1
boto3~=1.21.32
gensim~=4.1.2
botocore~=1.24.32
wheel~=0.37.1
importlib_metadata~=4.11.3
pip~=23.2.1
pytest~=7.1.1
joblib~=1.2.0
python~=3.9.12
pyzmq~=22.3.0
ray~=2.7.0
sqlalchemy~=1.4.32
pulsar~=2.10.4
google-protobuf~=4.21.12

